+++ app/javascript/components/shared/BillingDetail.vue
@@ -86,7 +86,12 @@
   import { ref, inject, onMounted, computed } from 'vue'
   import jwtFetch from '../../packs/jwtFetch'
   import { useCookies } from 'vue3-cookies'
-  import { getCurrentTime, getFirstDayOfTime, formatDate } from '../../packs/datetimeUtils'
+  import {
+    getCurrentTime,
+    getFirstDayOfTime,
+    formatDate,
+    isFutureDate
+  } from '../../packs/datetimeUtils'
 
   const props = defineProps({
     type: String,
@@ -127,15 +132,9 @@
   })
 
   const disabledDate = (date) => {
-    if (date) {
-      const current = new Date()
-      const year = date.getFullYear()
-      const month = date.getMonth()
-      const compareDate = new Date(year, month, 1)
-      return compareDate > current
-    }
+    return isFutureDate(date)
   }
-  
+
   const dateChange = (e) => {
     const dateString = formatDate(e) + ' 00:00:00'
     startTime.value = dateString
